move out helper methods from shared examples
helper methods should be made available through a module, not the shared
examples. These are run in a different context and so the helper method
won’t be available to other example groups. By moving the helper into a
module and including it for every controller, the asset specs are
mostly fixed