use index files for asset pipeline in Engine
It’s more easy to include index files in applications using this Engine.
The manifests for the assets is now compiled from the index files
describing which files to include. This allows to do something like:

```
//= require fassets_core
```

instead of

```
//= require fassets_core/application
```

as it had to be done, before this change. See the dummy application on
an easy example of how to use this.