add shared examples for ActsAsAsset
This brings new test helpers for new asset types. This describes, how a
type should behave like, when it acts as (an) asset. This can be used to
extend specs for new asset types to check for the generic assets
behaviour.

You can check ```spec/models/acts_as_asset_spec.rb``` for an example of
how to use the shared examples.